gnhf #19: Centralized CLI JSON read-with-default behavior in the shared atomic file utility and reused it without changing runtime behavior.
